Return a subarray when linear indexing with nonscalars
Since the returned array might have a single dimension that spans *across* multiple dimensions from the original array, there is no way that we could construct an AxisArray that would represent this data.  Either we need to return a plain-old-subarray or completely disallow the operation entirely and throw an error.