First pass at implementing axis indexing with traits
This is missing tests, but it's a start and works in my quick interactive tests.  This creates three AxisTypes: Categorical, Dimensional, and Unsupported, as determined by the axistype function.

I created the checkaxis function (but don't use it yet) to enforce type-specific invariants.

The axisindexes function is used to 'lower' the fancy indexing behavior to a supported basic indexing type (Int, Range, etc).

A fallback `getindex(A, ::Any...)` calls the axisindexes functions for each fancy indexing dimension.  It just works with the other behaviors (like `A[Axis{:col}(Interval(.1,.5))]`).