Major refactor: store axis info as an Axis tuple
Instead of using separate tuples of Symbols and Vectors, store them together in
a tuple of Axis objects.

Simplify constructors to accept Axis types for both dimension names and vectors at the same time.  Alternatively, just names or just vectors may be provided, with defaults filled in for the other.

AxisArrays now always have default Axis objects for all dimensions, which seems to simplify user-facing code.

Add a specialized similar method to return a preallocated AxisArray with a modified dimension for a particular Axis.