Only permit indexing of dimensional axes with non-Real and Intervals
The line between Integers and Reals is far too thin to perform dramatically different behaviors.  This simplifies things quite a bit, while retaining the bulk of the power.  An Interval of Integers is more cleanly represented by UnitRanges for direct indexing.  The only functionality we lose is windows of fixed number of indices about axis values... but that may be possible to shoe-horn into the `<|` (or whatever we end up choosing) operator later.