WIP: use IndexAxis for reductions
This _almost_ works, but it's not type stable. The trouble is that the
`IndexAxis` type _sometimes_ wants to know the exact Axis information (like
when used in `similar`), but other times it only cares about the Axis name
(like when choosing the dimensions for a reduction).  Specifically, this branch
runs into trouble with this base method:

```
    function reduced_indices(inds::Indices{N}, d::Int, rd::AbstractUnitRange) where N
        d < 1 && throw(ArgumentError("dimension must be ≥ 1, got $d"))
        if d == 1
            return (oftype(inds[1], rd), tail(inds)...)
        elseif 1 < d <= N
            return tuple(inds[1:d-1]..., oftype(inds[d], rd), inds[d+1:N]...)::typeof(inds)
```

I've broken the contract that `oftype`—that is, `convert`—returns an object
of exactly the requested type.  But here I simply want a new IndexAxis object
that has the same name and wraps the given `rd` range.