RFC: Re-enable axis checks in constructor
This is a major breaking change. Back before this package was officially registered, these checks were dropped. That means that many of the indexing methods that require sorted vectors are now broken. Enforcing these preconditions in the constructor is, unfortunately, terribly breaking.

I think the alternative here would be to delay the checks until indexing occurs, but the downside there is that for non-range axes, it would require a `issorted` pass through the data upon every operation. We could defer to the SortedVector type more, but that easily walks us into a type-unstable API. Thoughts?