Moved axisnames earlier in core.jl
In order to include AxisArrays in system images we needed to move the
`axisnames` definitions earlier in the core.jl file. This appears to be an issue
with generated functions and system images. In this case, we're calling the
generated similar function which reference axisdim outside of the quote block, which
then referenced the axisnames methods defined later in the file.