Try converting indexes once and then defer to the parent
AxisArrays has to handle many different index types, but it can't
handle types defined in other packages. This provides for a fallback
mode in which AxisArrays tries to convert "fancy" index types into
"regular" ones, but if this fails then it assumes this is something
the parent knows how to handle.